raspberry_pi_reading/master_raspberry_pi/requirements.txt
modbus_tk
pyserial
